wip(healthInfo): content update (#302)
* wip: content update

* feat(guide): add redirect link for healthInfo page

* refactor: add embed google maps
