+++ client/src/assets/placeholder/healthInfo.js
@@ -2,7 +2,7 @@
 export const INFO = Object.freeze({
   dropdown: [
     {
-      title: 'Specialists',
+      title: 'Doctors',
       content: [
         {
           name: 'DR. C.BHATTACHARYA',
@@ -10,7 +10,6 @@ export const INFO = Object.freeze({
           phone: '8895298094',
         },
         {
-          // year: 'Second Year',
           name: 'DR. S.R. BEHERA',
           room: 'VK Hall, B/103',
           phone: '7008485946',
@@ -33,7 +32,7 @@ export const INFO = Object.freeze({
       ],
     },
     {
-      title: 'Doctors',
+      title: 'Visiting Doctors',
       content: [
         {
           name: 'DR. S.N. MOHAPATRA',
@@ -57,14 +56,17 @@ export const INFO = Object.freeze({
       content: [
         {
           name: 'CWS Hospital',
+          address: 'Northern Ave, NIT Rourkela, Jagda, Rourkela, Odisha 769042',
           phone: '6612473931',
         },
         {
-          name: 'HI-TECH',
+          name: 'Hi-Tech Hospital',
+          address: 'Near Hanuman Vatika R.G.H Campus, Rourkela, Odisha 769004',
           phone: '6612400751',
         },
         {
           name: 'Lifeline Hsopital',
+          address: 'D/10, Civil Township, Rourkela, Odisha 769004',
           phone: '9437079322',
         },
       ],
@@ -73,28 +75,32 @@ export const INFO = Object.freeze({
       title: 'Ambulance',
       content: [
         {
-          name: 'NIT Ambulance',
+          services: 'NIT Rourkela Ambulance',
           phone: '8280468609',
         },
       ],
     },
   ],
 
-  location: [
-    {
+  location: {
+    health: {
       title: 'Health Centre',
       location1: 'NIT Rourkela ',
       location2: 'Sector-1, Rourkela ',
       location3: 'Odisha, India',
       pin: ' PIN: 769008',
+      imageAlt: 'Health Centre',
+      url: 'https://goo.gl/maps/Bcx8J9jpHXd2wy3D9',
     },
-    {
+    cws: {
       title: 'CWS Hospital',
       location1: 'Northern Ave, ',
       location2: 'NIT Rourkela ',
       location3: 'Jagda, Rourkela ',
       location4: 'Odisha, India',
       pin: ' PIN: 769042',
+      imageAlt: 'CWS Hospital',
+      url: 'https://goo.gl/maps/q27fmxb92qvJREhp8',
     },
-  ],
+  },
 });
