+++ client/src/components/guide/Options.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import Link from 'next/link';
 
 import makeStyles from '@mui/styles/makeStyles';
 import { Typography, Grid } from '@mui/material';
@@ -65,15 +66,17 @@ const Options = () => {
           <i className='fas fa-chevron-right'></i>
         </Typography>
       </div>
-      <div className={classes.optionWrapper}>
-        <Image src={icon4} alt='Icon' className={classes.icons} />
-        <Typography variant='h3' className={classes.option2}>
-          Health and Emergency Info
-        </Typography>
-        <Typography variant='h3' className={classes.arrow}>
-          <i className='fas fa-chevron-right'></i>
-        </Typography>
-      </div>
+      <Link href='/info/health' passHref>
+        <div className={classes.optionWrapper}>
+          <Image src={icon4} alt='Icon' className={classes.icons} />
+          <Typography variant='h3' className={classes.option2}>
+            Health and Emergency Info
+          </Typography>
+          <Typography variant='h3' className={classes.arrow}>
+            <i className='fas fa-chevron-right'></i>
+          </Typography>
+        </div>
+      </Link>
     </div>
   );
 };
