+++ client/src/pages/info/health.jsx
@@ -2,8 +2,8 @@ import React from 'react';
 import Head from 'next/head';
 
 // Components
-import Marginals from '../components/marginals/Marginals';
-import HealthAndEmergency from '../screens/HealthAndEmergency';
+import Marginals from '../../components/marginals/Marginals';
+import HealthInfo from '../../screens/HealthInfo';
 
 function HealthAndEmergencyInfo() {
   return (
@@ -74,7 +74,7 @@ function HealthAndEmergencyInfo() {
         />
       </Head>
       <Marginals>
-        <HealthAndEmergency />
+        <HealthInfo />
       </Marginals>
     </>
   );
